backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
youtube-transcript-api==0.6.2
google-generativeai==0.7.2
pydantic==2.7.4
python-dotenv==1.0.1
